CLI testing: Support module and e2e tests
