--- cli_tools_e2e_test/go.mod
@@ -1,17 +1,0 @@
-module github.com/GoogleCloudPlatform/compute-image-tools/cli_tools_e2e_test
-
-go 1.13
-
-require (
-	cloud.google.com/go/storage v1.10.0
-	github.com/GoogleCloudPlatform/compute-image-tools/cli_tools v0.0.0-20200813223603-3672ca27e050
-	github.com/GoogleCloudPlatform/compute-image-tools/daisy v0.0.0-20200813213118-5b8ac20eab97
-	github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils v0.0.0-20200813223603-3672ca27e050
-	github.com/aws/aws-sdk-go v1.34.4
-	golang.org/x/net v0.0.0-20200813134508-3edf25e44fcc // indirect
-	golang.org/x/tools v0.0.0-20200813203630-136574234359 // indirect
-	google.golang.org/api v0.30.0
-	google.golang.org/genproto v0.0.0-20200813001606-1ccf2a5ae4fd // indirect
-	google.golang.org/grpc v1.31.0 // indirect
-	honnef.co/go/tools v0.0.1-2020.1.5 // indirect
-)
