+++ cli_tools_tests/common/compute/image.go
