+++ cli_tools_tests/common/compute/instance.go
