+++ cli_tools_tests/common/utils/utils.go
