+++ cli_tools_tests/e2e/gce_image_import_export/test_suites/scripts/post_translate_test.wf.json
