+++ cli_tools_tests/e2e/gce_ovf_import/scripts/ovf_import_test_ubuntu_3_disks.sh
