+++ cli_tools_tests/e2e/gce_ovf_import/scripts/ovf_import_test_windows_two_disks.ps1
