+++ gce_ovf_import_tests.Dockerfile
@@ -16,15 +16,13 @@ FROM golang:alpine
 RUN apk add --no-cache git
 
 # Build test runner
-WORKDIR /cli_tools_e2e_test
-COPY cli_tools_e2e_test/ .
-RUN cd gce_ovf_import && CGO_ENABLED=0 go build -o /gce_ovf_import_test_runner
+COPY / /build
+WORKDIR /build
+RUN cd /build/cli_tools_tests/e2e/gce_ovf_import && CGO_ENABLED=0 go build -o /gce_ovf_import_test_runner
 RUN chmod +x /gce_ovf_import_test_runner
 
 # Build binaries to test
-WORKDIR /cli_tools
-COPY cli_tools/ .
-RUN cd gce_ovf_import && CGO_ENABLED=0 go build -o /gce_ovf_import
+RUN cd /build/cli_tools/gce_ovf_import && CGO_ENABLED=0 go build -o /gce_ovf_import
 RUN chmod +x /gce_ovf_import
 
 # Build test container
@@ -33,7 +31,7 @@ ENV GOOGLE_APPLICATION_CREDENTIALS /etc/compute-image-tools-test-service-account
 COPY --from=0 /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/ca-certificates.crt
 COPY --from=0 /gce_ovf_import_test_runner gce_ovf_import_test_runner
 COPY --from=0 /gce_ovf_import gce_ovf_import
-COPY /cli_tools_e2e_test/gce_ovf_import/scripts/ /scripts/
+COPY /cli_tools_tests/e2e/gce_ovf_import/scripts/ /scripts/
 COPY /daisy_integration_tests/scripts/ /daisy_integration_tests/scripts/
 COPY /daisy_workflows/ /daisy_workflows/
 ENTRYPOINT ["./wrapper", "./gce_ovf_import_test_runner"]
