+++ test-infra/prowjobs/gocheck/main.sh
@@ -33,8 +33,8 @@ TARGETS=("daisy"
          "cli_tools"
          "test-infra/prowjobs/cleanerupper"
          "test-infra/prowjobs/wrapper"
-         "cli_tools_e2e_test/gce_image_import_export"
-         "cli_tools_e2e_test/gce_ovf_import")
+         "cli_tools_tests/e2e/gce_image_import_export"
+         "cli_tools_tests/e2e/gce_ovf_import")
 for TARGET in "${TARGETS[@]}"; do
   echo "Checking ${TARGET}"
   cd "/${REPO_PATH}/${TARGET}"
